move i2cerror to shared module
