+++ shared/src/lib.rs
@@ -6,3 +6,4 @@ pub mod interrupts;
 pub mod kernel;
 pub mod sync;
 pub mod sys;
+pub mod i2c;
